DRF/requirements.txt
Django>=6.0,<7
djangorestframework>=3.14
django-filter>=23.0
